edit(lab2): code complete
